+++ UIViews/src/main/java/com/programmersbox/uiviews/presentation/NavGraph.kt
@@ -47,7 +47,7 @@ import com.programmersbox.uiviews.presentation.settings.extensions.ExtensionList
 import com.programmersbox.uiviews.presentation.settings.extensions.IncognitoScreen
 import com.programmersbox.uiviews.presentation.settings.moresettings.MoreSettingsScreen
 import com.programmersbox.uiviews.presentation.settings.updateprerelease.PrereleaseScreen
-import com.programmersbox.uiviews.presentation.webview.WebViewScreen
+import com.programmersbox.kmpuiviews.presentation.webview.WebViewScreen
 import com.programmersbox.uiviews.utils.NotificationLogo
 import com.programmersbox.uiviews.utils.sharedelements.animatedScopeComposable
 import com.programmersbox.uiviews.utils.trackScreen
