+++ gradle/libs.versions.toml
@@ -1,7 +1,7 @@
 [versions]
 biometricVersion = "1.4.0-alpha03"
 blurhash = "0.4.0-SNAPSHOT"
-composeWebview = "0.33.6"
+composeWebviewMultiplatform = "1.9.40"
 corePerformance = "1.0.0"
 generativeai = "0.9.0"
 gradle = "8.9.2"
@@ -121,7 +121,7 @@ androidx-core-performance = { module = "androidx.core:core-performance", version
 biometric = { module = "androidx.biometric:biometric", version.ref = "biometricVersion" }
 blurhash = { module = "com.vanniktech:blurhash", version.ref = "blurhash" }
 compose-collapsable = { module = "me.tatarka.compose.collapsable:compose-collapsable", version.ref = "composeCollapsable" }
-compose-webview = { module = "io.github.kevinnzou:compose-webview", version.ref = "composeWebview" }
+compose-webview-multiplatform = { module = "io.github.kevinnzou:compose-webview-multiplatform", version.ref = "composeWebviewMultiplatform" }
 dragselect = { module = "com.dragselectcompose:dragselect", version.ref = "dragselect" }
 easylauncher = { module = "com.project.starter:easylauncher", version.ref = "easylauncher" }
 firebase-crashlytics-gradle = { module = "com.google.firebase:firebase-crashlytics-gradle", version.ref = "firebaseCrashlyticsGradle" }
