+++ kmpuiviews/src/commonMain/kotlin/com/programmersbox/kmpuiviews/presentation/webview/WebViewScreen.kt
@@ -1,5 +1,6 @@
-package com.programmersbox.uiviews.presentation.webview
+package com.programmersbox.kmpuiviews.presentation.webview
 
+import androidx.compose.foundation.basicMarquee
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -7,11 +8,12 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
 import androidx.compose.ui.Modifier
-import com.kevinnzou.web.WebView
-import com.kevinnzou.web.rememberWebViewState
+import com.multiplatform.webview.web.WebView
+import com.multiplatform.webview.web.rememberWebViewState
 import com.programmersbox.kmpuiviews.presentation.components.BackButton
-import com.programmersbox.uiviews.utils.HideNavBarWhileOnScreen
+import com.programmersbox.kmpuiviews.utils.HideNavBarWhileOnScreen
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -20,17 +22,33 @@ fun WebViewScreen(
 ) {
     HideNavBarWhileOnScreen()
     val state = rememberWebViewState(url = url)
+
+    DisposableEffect(Unit) {
+        state.webSettings.apply {
+            isJavaScriptEnabled = true
+            androidWebSettings.apply {
+                isAlgorithmicDarkeningAllowed = true
+                safeBrowsingEnabled = true
+            }
+        }
+        onDispose { }
+    }
+
     Scaffold(
         topBar = {
             TopAppBar(
-                title = { Text(url) },
+                title = {
+                    Text(
+                        url,
+                        modifier = Modifier.basicMarquee()
+                    )
+                },
                 navigationIcon = { BackButton() }
             )
         }
     ) { padding ->
         WebView(
             state = state,
-            onCreated = { it.settings.javaScriptEnabled = true },
             modifier = Modifier
                 .fillMaxSize()
                 .padding(padding)
